voice-agent/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Speech (Google Cloud)
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.16.1

# LLM
google-generativeai==0.4.0
openai==1.12.0

# Audio processing
soundfile==0.12.1
numpy==1.26.3

# Logging & Monitoring
structlog==24.1.0

# Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0

# Testing
pytest==7.2.0
pytest-asyncio==0.23.4
httpx==0.26.0
